Codex: Remove Live Room AI assistant. v5.15.24
Delete the Live Room assistant UI, conversation backend, popout routes,
styles, greeting audio, and exclusive translations. Preserve Live Room
streaming and its roomToken API field, and move shared AI configuration
types to openai.go without changing OCR or dubbing behavior.

Bump the platform version to v5.15.24 and update the changelog. Leave tests
unchanged. Verified all eight local suites, 11 UI tests, Go checks, lint,
and English/Chinese dashboard builds. External AI tests were not run.

---------
