+++ DEVELOPER.md
@@ -1287,6 +1287,7 @@ The following are the update records for the Oryx server.
     * Dubbing: Support space key to play/pause. v5.15.21
     * AI: Support OpenAI o1-preview model. v5.15.22
     * AI: Move Transcript, Dubbing, and OCR scenarios to discouraged. v5.15.23
+    * Live Room: Remove the AI assistant. v5.15.24
 * v5.14:
     * Merge features and bugfix from releases. v5.14.1
     * Dubbing: Support VoD dubbing for multiple languages. [v5.14.2](https://github.com/ossrs/oryx/releases/tag/v5.14.2)
