+++ platform/main.go
@@ -215,10 +215,6 @@ func doMain(ctx context.Context) error {
 		return errors.Wrapf(err, "start OCR worker")
 	}
 
-	// Create AI Talk worker for live room.
-	talkServer = NewTalkServer()
-	defer talkServer.Close()
-
 	// Create AI Dubbing server for VoD translation.
 	dubbingServer = NewDubbingServer()
 	defer dubbingServer.Close()
@@ -437,7 +433,7 @@ func initPlatform(ctx context.Context) error {
 		"containers/data/dvr", "containers/data/record", "containers/data/vod",
 		"containers/data/upload", "containers/data/vlive", "containers/data/signals",
 		"containers/data/lego", "containers/data/.well-known", "containers/data/config",
-		"containers/data/transcript", "containers/data/srs-s3-bucket", "containers/data/ai-talk",
+		"containers/data/transcript", "containers/data/srs-s3-bucket",
 		"containers/data/dubbing", "containers/data/ocr",
 	} {
 		if _, err := os.Stat(dir); err != nil && os.IsNotExist(err) {
