+++ platform/service.go
@@ -253,10 +253,6 @@ func handleHTTPService(ctx context.Context, handler *http.ServeMux) error {
 		return errors.Wrapf(err, "handle dubbing")
 	}
 
-	if err := handleAITalkService(ctx, handler); err != nil {
-		return errors.Wrapf(err, "handle AI talk")
-	}
-
 	var ep string
 
 	handleHostVersions(ctx, handler)
