+++ ui/src/resources/locale.json
@@ -489,12 +489,6 @@
         "room": {
           "nav": "导航",
           "stream": "直播流",
-          "aic": "AI配置",
-          "enable": "启用AI助手",
-          "disable": "关闭AI助手",
-          "update": "更新助手",
-          "name": "AI助手名称",
-          "name2": "(可选) 设置AI助手的名称",
           "provider": "AI服务商",
           "provider2": "请选择AI服务商",
           "model": "AI模型",
@@ -505,26 +499,8 @@
           "window2": "设置AI上下文消息的最大数目，越大体验越好但成本越高",
           "words": "AI响应长度",
           "words2": "设置AI响应的最大长度，越大体验越好但成本越高",
-          "aiw": "AI助手",
-          "aiwe": "请先配置AI助手",
-          "https": "请使用HTTPS访问管理后台。若使用自签名证书，请点击高级，然后继续访问。",
-          "mic": "访问麦克风失败",
-          "talk": "开始聊天",
-          "speaker": "扬声器播放音频失败",
-          "usage": "按住说话，松开结束，和微信语音一样",
-          "usage2": "按住空格键或者R键说话，松开结束",
-          "popout": "新的消息将展示在这里，请等待",
-          "ait": "直播间AI助手",
-          "ait2": "聊天消息",
-          "popchat": "弹出聊天窗口",
-          "popchat2": "弹出AI助手",
-          "settings": "设置会话",
-          "uname": "用户名",
-          "uname2": "(可选) 设置用户名，多人聊天可以方便的区分用户",
-          "basic": "基本设置",
           "asr": "ASR语音转文字",
           "chat": "AI模型配置",
-          "tts": "TTS文字转语音",
           "asre": "开启ASR",
           "chate": "开启AI处理",
           "ttse": "开启TTS",
@@ -532,17 +508,7 @@
           "shdr": "显示时间",
           "sasr": "显示原文",
           "strans": "显示翻译",
-          "text": "请输入聊天文字",
-          "assistant": "助手链接",
-          "pattern": "AI工作模式",
-          "pattern2": "请选择AI的工作模式，比如是对话还是听写",
-          "patternChat": "AI和用户对话方式，一来一回的对话",
-          "patternListen": "AI一直听用户对话，用户一直说，听写模式",
-          "startd": "开始听写",
-          "asrp": "ASR提示词",
-          "asrp2": "请选择生成ASR提示词的方式",
-          "rbasic": "房间基本设置",
-          "post": "后处理"
+          "rbasic": "房间基本设置"
         }
       },
       "plat": {
@@ -831,17 +797,7 @@
       },
       "lr": {
         "room": {
-          "post": "Post Processing",
           "rbasic": "Room Settings",
-          "asrp2": "Please select the way to generate ASR prompt",
-          "asrp": "ASR Prompt",
-          "startd": "Start Dictation",
-          "patternListen": "AI listens to the user speech, and the user keeps talking",
-          "patternChat": "AI and user chat, taking turns with each sentence.",
-          "pattern2": "Please select AI work pattern, e.g. chat or dictation",
-          "pattern": "AI Pattern",
-          "assistant": "Assistant Link",
-          "text": "Please input chat text",
           "ttse": "Enable TTS",
           "fse": "Fullscreen",
           "shdr": "Show Time",
@@ -849,26 +805,8 @@
           "strans": "Show Translation",
           "chate": "Enable AI processing",
           "asre": "Enable ASR",
-          "tts": "TTS Text to Speech",
           "chat": "AI Instructions",
           "asr": "ASR Speech to Text",
-          "basic": "Basic Settings",
-          "uname2": "(Optional) Setup user name, for multi-user chat",
-          "uname": "User Name",
-          "settings": "Session Settings",
-          "popchat2": "Popout AI Assistant",
-          "popchat": "Popout Live Chat",
-          "ait2": "Live Chat",
-          "ait": "AI Assistant",
-          "popout": "New messages will be displayed here, please wait",
-          "usage2": "Press the R key or SPACE to talk",
-          "usage": "Press and hold to talk, release to end",
-          "speaker": "Failed to play audio",
-          "talk": "Start Chatting",
-          "mic": "Open microphone failed",
-          "https": "Please use HTTPS to access the management background. If you use a self-signed certificate, please click Advanced, and then continue to access.",
-          "aiwe": "Please setup AI assistant first",
-          "aiw": "AI Assistant",
           "words2": "Setup the max number of AI response words, the more the better experience but the higher the cost",
           "words": "AI Response Words",
           "window2": "Setup the max number of AI context messages, the more the better experience but the higher the cost",
@@ -879,12 +817,6 @@
           "model": "AI Model Name",
           "provider2": "Please select AI service provider",
           "provider": "AI Provider",
-          "name2": "(Optional) Setup AI assistant name",
-          "name": "AI Assistant Name",
-          "update": "Update AI Assistant",
-          "disable": "Disable AI Assistant",
-          "enable": "Enable AI Assistant",
-          "aic": "AI Configuration",
           "stream": "Live Stream",
           "nav": "Navigation"
         },
